requirements.txt
# Core Dependencies
watchdog>=2.4.1
GitPython>=3.1.43

# Web UI for Review Queue
fastapi>=0.111.0
uvicorn>=0.30.1
Jinja2>=3.1.4
python-multipart>=0.0.9
aiofiles>=23.2.1
aiosqlite>=0.20.0
PyYAML>=6.0.1